+++ assets/javascripts/kitten/components/form/rich-text-input.test.js
@@ -13,9 +13,7 @@ describe('<RichTextInput />', () => {
   describe('with config prop', () => {
     it('pass the config prop to <RichTextInput />', () => {
       const locale = { locale: 'fr' }
-      const component = shallow(
-        <RichTextInput config={ locale } />
-      )
+      const component = shallow(<RichTextInput config={locale} />)
       const expectedProps = { config: locale }
 
       expect(component.props()).toMatchObject(expectedProps)
@@ -24,16 +22,13 @@ describe('<RichTextInput />', () => {
 
   describe('with toolbar prop', () => {
     it('pass the toolbar prop to <RichTextInput />', () => {
-      const toolbar =
-        [{
+      const toolbar = [
+        {
           name: 'basicstyles',
-          items: [
-            'Bold', 'Italic', 'Underline', 'Link', 'Image',
-          ],
-        }]
-      const component = shallow(
-        <RichTextInput toolbar={ toolbar } />
-      )
+          items: ['Bold', 'Italic', 'Underline', 'Link', 'Image'],
+        },
+      ]
+      const component = shallow(<RichTextInput toolbar={toolbar} />)
       const expectedProps = { toolbar: toolbar }
 
       expect(component.props().config).toMatchObject(expectedProps)