+++ assets/javascripts/kitten/components/form/switch.js
@@ -3,22 +3,22 @@ import classNames from 'classnames'
 
 export class Switch extends React.Component {
   render() {
-    let switchClassNames = classNames(
-      'k-Switch__label',
-      { 'k-Switch__label--big': this.props.big },
-    )
+    let switchClassNames = classNames('k-Switch__label', {
+      'k-Switch__label--big': this.props.big,
+    })
 
     return (
       <div className="k-Switch">
-        <input id={ this.props.id }
-                type="checkbox"
-                className="k-Switch__input"
-                defaultChecked={ this.props.isChecked }
-                disabled={ this.props.disabled } />
+        <input
+          id={this.props.id}
+          type="checkbox"
+          className="k-Switch__input"
+          defaultChecked={this.props.isChecked}
+          disabled={this.props.disabled}
+        />
 
-        <label htmlFor={ this.props.id }
-               className={ switchClassNames } >
-          { this.props.label }
+        <label htmlFor={this.props.id} className={switchClassNames}>
+          {this.props.label}
         </label>
       </div>
     )