+++ assets/javascripts/kitten/components/meters/progress.js
@@ -5,18 +5,9 @@ const valueMax = 100
 
 export class Progress extends Component {
   render() {
-    const {
-      className,
-      color,
-      value,
-      rampProps,
-      ...others,
-    } = this.props
+    const { className, color, value, rampProps, ...others } = this.props
 
-    const progressClassName = classNames(
-      'k-Progress',
-      className,
-    )
+    const progressClassName = classNames('k-Progress', className)
 
     const progressValue = value > valueMax ? valueMax : value
 
@@ -25,24 +16,19 @@ export class Progress extends Component {
       width: `${progressValue}%`,
     }
 
-    const rampClassName = classNames(
-      'k-Progress__ramp',
-      rampProps.className,
-    )
+    const rampClassName = classNames('k-Progress__ramp', rampProps.className)
 
     return (
       <div
-        { ...others }
+        {...others}
         role="progressbar"
         aria-valuemin="0"
-        aria-valuemax={ `${valueMax}` }
-        aria-valuenow={ progressValue }
-        className={ progressClassName }>
-        <div { ...rampProps } className={ rampClassName }>
-          <div
-            className="k-Progress__slider"
-            style={ style }
-          />
+        aria-valuemax={`${valueMax}`}
+        aria-valuenow={progressValue}
+        className={progressClassName}
+      >
+        <div {...rampProps} className={rampClassName}>
+          <div className="k-Progress__slider" style={style} />
         </div>
       </div>
     )