+++ assets/javascripts/kitten/components/typography/title.test.js
@@ -2,8 +2,7 @@ import React from 'react'
 import { Title } from 'kitten/components/typography/title'
 
 describe('Title with default props', () => {
-  const defaultComponent = shallow(
-    <Title />)
+  const defaultComponent = shallow(<Title />)
 
   it('has a default title attribute', () => {
     expect(defaultComponent.find('.k-Title')).toHaveLength(1)
@@ -18,9 +17,7 @@ describe('Title with default props', () => {
   })
 
   describe('<Title />', () => {
-    const component = shallow(
-      <Title className="k-Title--custom">Felis…</Title>
-    )
+    const component = shallow(<Title className="k-Title--custom">Felis…</Title>)
 
     it('renders a <Title class="k-Title" />', () => {
       expect(component.type()).toBe('h1')
@@ -31,24 +28,20 @@ describe('Title with default props', () => {
 
     describe('modifier', () => {
       it('accepts secondary', () => {
-        const component = shallow(
-          <Title modifier="secondary" />
-        )
+        const component = shallow(<Title modifier="secondary" />)
         expect(component.hasClass('k-Title--secondary')).toBe(true)
       })
     })
 
     describe('with custom tag', () => {
       it('renders h2', () => {
-        const component = shallow(
-          <Title tag="h2" />
-        )
+        const component = shallow(<Title tag="h2" />)
         expect(component.type()).toBe('h2')
       })
     })
 
     describe('with margin prop', () => {
-      const component = shallow(<Title margin={ false } />)
+      const component = shallow(<Title margin={false} />)
 
       it('has a good class', () => {
         expect(component.hasClass('k-Title--withoutMargin')).toBe(true)