+++ assets/javascripts/kitten/components/uploaders/uploader.test.js
@@ -40,12 +40,12 @@ describe('<Uploader />', () => {
   describe('with theme prop', () => {
     const component = mount(
       <Uploader
-        theme={ CustomTheme }
+        theme={CustomTheme}
         buttonLabel="Custom label"
         disabled
         fileName="custom-filename.jpg"
-        deletable={ false }
-      />
+        deletable={false}
+      />,
     )
     const theme = component.find(CustomTheme).first()