+++ assets/javascripts/kitten/helpers/dom/element-helper.js
@@ -56,8 +56,11 @@ export const domElementHelper = {
    * Return if DOM API can be used.
    */
   canUseDom() {
-    return !!(typeof window !== 'undefined' &&
-              window.document && window.document.createElement)
+    return !!(
+      typeof window !== 'undefined' &&
+      window.document &&
+      window.document.createElement
+    )
   },
 }