+++ assets/javascripts/kitten/helpers/utils/string.js
@@ -3,9 +3,8 @@ export const stringUtils = {
    * Converts the first character to uppercase.
    */
   upcaseFirst(str) {
-    if (!str || typeof(str) !== 'string')
-      return str
+    if (!str || typeof str !== 'string') return str
 
     return str.charAt(0).toUpperCase() + str.slice(1)
-  }
+  },
 }