+++ CHANGELOG.md
@@ -5,7 +5,16 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ## [unreleased]
 
 - Feature: Update A11Y on `Modal` component.
-- Feature: Add `subTitleTruncate` prop for `CrowdfundingCard`.
+
+## [20.6.0] - 2018-04-06
+
+Features:
+- Add `subTitleTruncate` prop for `CrowdfundingCard`.
+- Add `horizontalStroke` prop for `SimpleCard`.
+
+Fixes:
+- Update `:hover` and `:focus` state on `azote` button.
+- Fix `CrowdfundingCard` spacing.
 
 ## [20.5.0] - 2018-04-03